Fixed typescript to actually work
